fix a problem with java-dialogs which dont open correct (thanx Scott Moser)
furthermore, and this is really interesting, this seems to fix also the bug
widely known as g.o.d - the grab of death. the g.o.d seems to freeze fluxbox
but let the mouse still be moving and no high-cpuload can be seen. the
(un)lucky can force the g.o.d by clicking like a weirdo and pressing keys like
crazy and then maybe it will happen. we are not quite sure WHY exactly this
happens but as it seems: XSync(true) caused the trouble.

we still keep watching this issue but hopefully noone experiences bad
"freezes" anymore. (/me crosses fingers)