build-sys: use iconv library detections from gettext
This patch took copy of config.rpath from gettext 0.18.2 to build-aux/
directory.  Software projects usually does not need to do that, because
the file is copied by gettext autopoint in autogen.sh or bootstrap.

In Fluxbox case calling autopoint seems wrong, as it would generate po/
directory with necessary make automation, adn 'msgid' & 'msgstr' files,
but that is not how the translations are done in this project.