Initial implementation of shortcut to windows
[PURPOSE]

In editors such as vi and emacs, a user can mark a line in a file with
a shortcut key and afterwards jump back to that line using the shortcut.

The idea is extended to opened windows. A user can assign a keyboard
shortcut to an opened window.  Afterwards, the shortcut can be used to
switch focus back to the marked window.

Such shortcuts save the user from pressing "alt+tab" multiple times to
cycle through windows until the desired one is found.

[EXAMPLE USAGE]

The following binding is added to file "~/.fluxbox/keys":

    Mod1 m ARG :MarkWindow

    Mod1 g ARG :GotoMarkedWindow

User enters "alt+m x" to mark the currently focused window with shortcut
key 'x'

User enters "alt+g x" to switch focus to the marked window

[IMPLEMENTATION SUMMARY]

- Two new commands were added :MarkWindow and :GotoMarkedWindow

- Keys.cc was modified:
    - addBinding() method supports parsing an argument placeholder where
      the user can pass in a shortcut key
    - doAction() method forwards the shortcut key to the command to execute
    - Class Keys::t_key was modified to recognize a placeholder key

- New class ShortcutManager was added to maintain mapping of shortcut
  keys to marked windows