Disable SIGALRM before launching other WMs
Via 43bdf499, `alarm(2)` might get issued which might trigger a SIGALRM
in some edge cases in a shutdown / restart scenario. In case of a
restart-to-other-WM scenario, this will yield a SIGALRM in the other WM
which then might fail because no SIGALRM handler is installed.

This commit addresses the issue by disabling any alarm to our own
instance before launching the other WM.